machine-learning-client/requirements.txt
flask
google-genai
python-dotenv
pymongo
pytest
pytest-cov
